+++ src/Symfony/Bundle/FrameworkBundle/Translation/TranslationLoader.php
@@ -42,12 +42,12 @@ class TranslationLoader implements TranslationReaderInterface
     }
 
     /**
-     * Loads translation messages from a directory to the catalogue.
+     * Reads translation messages from a directory to the catalogue.
      *
      * @param string           $directory the directory to look into
      * @param MessageCatalogue $catalogue the catalogue
      */
-    public function loadMessages($directory, MessageCatalogue $catalogue)
+    public function read($directory, MessageCatalogue $catalogue)
     {
         if (!is_dir($directory)) {
             return;
@@ -64,4 +64,18 @@ class TranslationLoader implements TranslationReaderInterface
             }
         }
     }
+
+    /**
+     * Loads translation messages from a directory to the catalogue.
+     *
+     * @param string           $directory the directory to look into
+     * @param MessageCatalogue $catalogue the catalogue
+     *
+     * @deprecated since 3.4 will be removed in 4.0. Use read instead.
+     */
+    public function loadMessages($directory, MessageCatalogue $catalogue)
+    {
+        $this->read($directory, $catalogue);
+    }
+
 }